+++ internal/testing/testdata.go
@@ -17,11 +17,13 @@ package testdata
 
 import (
 	"context"
+	_ "embed"
 	"encoding/base64"
 	"encoding/json"
+	"fmt"
 	"reflect"
 
-	"github.com/guacsec/guac/internal/testing/ingestor/keyutil"
+	"github.com/guacsec/guac/internal/testing/keyutil"
 	"github.com/guacsec/guac/pkg/assembler"
 	"github.com/guacsec/guac/pkg/handler/processor"
 	"github.com/guacsec/guac/pkg/ingestor/key"
@@ -30,6 +32,58 @@ import (
 )
 
 var (
+	// based off https://github.com/spdx/spdx-examples/blob/master/example7/spdx/example7-third-party-modules.spdx.json
+	//go:embed testdata/small-spdx.json
+	SpdxExampleSmall []byte
+
+	//go:embed testdata/alpine-spdx.json
+	SpdxExampleBig []byte
+
+	//go:embed testdata/alpine-small-spdx.json
+	SpdxExampleAlpine []byte
+
+	// Invalid types for field spdxVersion
+	//go:embed testdata/invalid-spdx.json
+	SpdxInvalidExample []byte
+
+	// Example scorecard
+	//go:embed testdata/kubernetes-scorecard.json
+	ScorecardExample []byte
+
+	// Invalid scorecard
+	//go:embed testdata/invalid-scorecard.json
+	ScorecardInvalid []byte
+
+	//go:embed testdata/alpine-cyclonedx.json
+	CycloneDXExampleAlpine []byte
+
+	//go:embed testdata/quarkus-deps-cyclonedx.json
+	CycloneDXExampleQuarkusDeps []byte
+
+	//go:embed testdata/small-deps-cyclonedx.json
+	CycloneDXExampleSmallDeps []byte
+
+	//go:embed testdata/invalid-cyclonedx.json
+	CycloneDXInvalidExample []byte
+
+	//go:embed testdata/distroless-cyclonedx.json
+	CycloneDXDistrolessExample []byte
+
+	//go:embed testdata/busybox-cyclonedx.json
+	CycloneDXBusyboxExample []byte
+
+	//go:embed testdata/big-mongo-cyclonedx.json
+	CycloneDXBigExample []byte
+
+	//go:embed testdata/crev-review.json
+	ITE6CREVExample []byte
+
+	//go:embed testdata/github-review.json
+	ITE6ReviewExample []byte
+
+	//go:embed testdata/certify-vuln.json
+	ITE6VulnExample []byte
+
 	// DSSE/SLSA Testdata
 
 	// Taken from: https://slsa.dev/provenance/v0.1#example
@@ -628,4 +682,91 @@ func GuacEdgeSliceEqual(slice1, slice2 []assembler.GuacEdge) bool {
 		}
 	}
 	return result
+}
+
+func existAndPop(nodes []*processor.DocumentNode, n *processor.DocumentNode) bool {
+	for i, nn := range nodes {
+		if docNodeEqual(nn, n) {
+			nodes = append(nodes[:i], nodes[i+1:]...) //nolint: staticcheck
+			return true
+		}
+	}
+	return false
+}
+
+func docEqual(a, b *processor.Document) bool {
+	a.Blob = ConsistentJsonBytes(a.Blob)
+	b.Blob = ConsistentJsonBytes(b.Blob)
+	return reflect.DeepEqual(a, b)
+}
+
+func DocTreeEqual(a, b processor.DocumentTree) bool {
+	return docNodeEqual(a, b)
+}
+
+func docNodeEqual(a, b *processor.DocumentNode) bool {
+	if a == nil || b == nil {
+		return false
+	}
+
+	// check if a and b Docuemnts are equal
+	if !docEqual(a.Document, b.Document) {
+		return false
+	}
+
+	// check if len of children are equal
+	if len(a.Children) != len(b.Children) {
+		return false
+	}
+
+	if len(a.Children) > 0 {
+		// Copy list of documentNodes of A
+		aCopy := make([]*processor.DocumentNode, len(a.Children))
+		copy(aCopy, a.Children)
+
+		// for each document in B, check exists and pop on listA
+		// where exists and pop equivalency
+		for _, bNode := range b.Children {
+			if !existAndPop(aCopy, bNode) {
+				return false
+			}
+		}
+	}
+
+	return true
+}
+
+// consistentJsonBytes makes sure that the blob byte comparison
+// does not differ due to whitespace in testing definitions.
+func ConsistentJsonBytes(b []byte) []byte {
+	var v interface{}
+	err := json.Unmarshal(b, &v)
+	if err != nil {
+		panic(err)
+	}
+	out, _ := json.Marshal(v)
+	return out
+}
+
+func StringTree(n *processor.DocumentNode) string {
+	return stringTreeHelper(n, "")
+}
+
+func stringTreeHelper(n *processor.DocumentNode, prefix string) string {
+	str := fmt.Sprintf("%s { doc: %s, %v, %v, %v}", prefix, string(ConsistentJsonBytes(n.Document.Blob)),
+		n.Document.Format,
+		n.Document.Type,
+		n.Document.SourceInformation,
+	)
+	for _, c := range n.Children {
+		str += "\n" + stringTreeHelper(c, prefix+"-")
+	}
+	return str
+}
+
+func DocNode(v *processor.Document, children ...*processor.DocumentNode) *processor.DocumentNode {
+	return &processor.DocumentNode{
+		Document: v,
+		Children: children,
+	}
 }
