+++ internal/testing/testdata/alpine-cyclonedx.json
