+++ internal/testing/testdata/alpine-small-spdx.json
