+++ internal/testing/testdata/alpine-spdx.json
