+++ internal/testing/testdata/big-mongo-cyclonedx.json
