+++ internal/testing/testdata/busybox-cyclonedx.json
