+++ internal/testing/testdata/distroless-cyclonedx.json
