+++ internal/testing/testdata/invalid-cyclonedx.json
