+++ internal/testing/testdata/invalid-scorecard.json
