+++ internal/testing/testdata/invalid-spdx.json
