+++ internal/testing/testdata/kubernetes-scorecard.json
