+++ internal/testing/testdata/quarkus-deps-cyclonedx.json
