+++ internal/testing/testdata/small-deps-cyclonedx.json
