+++ internal/testing/testdata/small-spdx.json
