+++ pkg/handler/processor/guesser/type_scorecard_test.go
@@ -18,7 +18,7 @@ package guesser
 import (
 	"testing"
 
-	testdata "github.com/guacsec/guac/internal/testing/processor"
+	testdata "github.com/guacsec/guac/internal/testing"
 	"github.com/guacsec/guac/pkg/handler/processor"
 )
