+++ pkg/handler/processor/ite6/ite6_test.go
@@ -19,7 +19,7 @@ import (
 	"reflect"
 	"testing"
 
-	testdata "github.com/guacsec/guac/internal/testing/processor"
+	testdata "github.com/guacsec/guac/internal/testing"
 	"github.com/guacsec/guac/pkg/handler/processor"
 )
