+++ pkg/ingestor/parser/cyclonedx/parser_cyclonedx_test.go
@@ -19,8 +19,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/guacsec/guac/internal/testing/ingestor/testdata"
-	processor_data "github.com/guacsec/guac/internal/testing/processor"
+	testdata "github.com/guacsec/guac/internal/testing"
 	"github.com/guacsec/guac/pkg/assembler"
 	"github.com/guacsec/guac/pkg/handler/processor"
 	"github.com/guacsec/guac/pkg/logging"
@@ -37,7 +36,7 @@ func Test_cyclonedxParser(t *testing.T) {
 	}{{
 		name: "valid small CycloneDX document",
 		doc: &processor.Document{
-			Blob:   processor_data.CycloneDXDistrolessExample,
+			Blob:   testdata.CycloneDXDistrolessExample,
 			Format: processor.FormatJSON,
 			Type:   processor.DocumentCycloneDX,
 			SourceInformation: processor.SourceInformation{
@@ -51,7 +50,7 @@ func Test_cyclonedxParser(t *testing.T) {
 	}, {
 		name: "valid small CycloneDX document with package dependencies",
 		doc: &processor.Document{
-			Blob:   processor_data.CycloneDXExampleSmallDeps,
+			Blob:   testdata.CycloneDXExampleSmallDeps,
 			Format: processor.FormatJSON,
 			Type:   processor.DocumentCycloneDX,
 			SourceInformation: processor.SourceInformation{
