+++ pkg/ingestor/parser/spdx/parse_spdx_test.go
@@ -19,8 +19,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/guacsec/guac/internal/testing/ingestor/testdata"
-	processor_data "github.com/guacsec/guac/internal/testing/processor"
+	testdata "github.com/guacsec/guac/internal/testing"
 	"github.com/guacsec/guac/pkg/assembler"
 	"github.com/guacsec/guac/pkg/handler/processor"
 	"github.com/guacsec/guac/pkg/logging"
@@ -37,7 +36,7 @@ func Test_spdxParser(t *testing.T) {
 	}{{
 		name: "valid big SPDX document",
 		doc: &processor.Document{
-			Blob:   processor_data.SpdxExampleAlpine,
+			Blob:   testdata.SpdxExampleAlpine,
 			Format: processor.FormatJSON,
 			Type:   processor.DocumentSPDX,
 			SourceInformation: processor.SourceInformation{
